Port: Ceph recovery logic improvements for blocked ops
Our current version of Jewel LTS (v10.2.6) has an issue on cluster
recovery where we will see blocked operations associated with an OSD
under certain conditions.

The cluster will remain in a HEALTH_WARN state and exhibit the following
warning:
       1 ops are blocked > 1048.58 sec on osd.1

This will block patching orchistration. This commit enhances the
recovery code so that we will look for this condition and restart the
OSD that sees these blocked operations.

From Wind River:
    CGTS-7216: Ceph recovery logic improvements for blocked ops
        0001____src_init-ceph.in.patch
        0002____wrs_ceph_pmon_wrapper.sh.patch

Signed-off-by: Robert Church <robert.church@windriver.com>
Signed-off-by: Changcheng Liu <changcheng.liu@intel.com>
Signed-off-by: Dehao Shang <dehao.shang@intel.com>