Port: wait for osd to enter some status
This is the ceph update part for this issue solution.

This update adds a new "osd-wait-status" script used to wait for one or a
group of OSDs to match one or a group of statuses as reported
by "ceph osd tree".

This change, with the start/stop services update to maintenance, blocks
maintenance from locking a storage host until the ceph OSD cluster reports
corresponding OSDs are down.

Inspection Record: http://crucible.wrs.com/cru/CR-CGCS-5068

From Wind River:
    CGTS-7174: CLONE - Upgrades: Upgrade strategy failed to upgrade storage node after locking
        0001____ceph.spec.patch
        0002____src_init-ceph.in.patch

Signed-off-by: Eric MacDonald <eric.macdonald@windriver.com>
Signed-off-by: Changcheng Liu <changcheng.liu@intel.com>
Signed-off-by: Dehao Shang <dehao.shang@intel.com>