Update verbose rados & rbd exceptions messages
Commit https://github.com/starlingx-staging/stx-ceph/commit/53e93f was
ported from Jewel to Mimic but was not updated for changes in make_ex()
for the named errno parameter. As a result we are seeing

TypeError: _init_() got an unexpected keyword argument 'errno'

This updates the code accordingly to support this and not raise an
exception.

Signed-off-by: Robert Church <robert.church@windriver.com>