Run xfs_repair if neeeded when activating osd
When doing B&R with multiple nodes with the intention of keeping osd
data, the osd data might be lost.
The filesystem on the osd becomes corrupted.

Recover using xfs_repair tool.
The filesystem has valuable metadata changes in a log which needs to
be replayed. Mount the filesystem to replay the log, then unmount it
before running xfs_repair.

The logic implemented must produce logs.
Use log level warning because level info is not shown in the output.

Closes-Bug: https://bugs.launchpad.net/starlingx/+bug/1910776
Signed-off-by: Dan Voiculeasa <dan.voiculeasa@windriver.com>