Add more OSD logging to the Ceph init script
In the https://bugs.launchpad.net/starlingx/+bug/1931103
the osd is stuck and we don't know the root cause for that.

In this commit we add more logging to the ceph init script to
report the state of the OSDs in order to have a better understanding
of the issue. Also, this will help us design recovery code to fix
this in the future.

Partial-Bug: 1931103
Signed-off-by: Mihnea Saracin <Mihnea.Saracin@windriver.com>