Fix ceph-disk output for multipath disks
When running ceph-disk on a system whithout multipath disks, the
output for Ceph OSD and Journal looks like:

/dev/sdb :
 /dev/sdb1 ceph data, prepared, cluster ceph, osd.0, osd uuid \
      e3c08a72-c755-4dec-b353-e4df4b4690c4, journal /dev/sdb2
 /dev/sdb2 ceph journal, for /dev/sdb1

And when running same command on a system with multipath disks, the
output is:

/dev/dm-5 :
 /dev/dm-6 data, xfs, mounted on /var/lib/ceph/osd/ceph-1
 /dev/dm-7 journal, 45b0969e-8ae0-4982-bf9d-5a8d867af560

With this fix, the output looks the same for system with and whithout
multipath disks.

This error was preventing the system to identify ceph OSDs were
already configured and ODSs were being zapped every host reboot.

Test Plan:
    PASS: Fresh install AIO-SX with multipath disks and
          Ceph backend
    PASS: Run ceph-disk command and verify output for osds

Closes-bug: 2009227

Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>