Fix execute_ceph_cmd function in ceph_common.sh
The function execute_ceph_cmd in ceph_common.sh was using the 'output'
variable as global and it was not being initialized before the first
use. This way, the CEPH_FAILURE variable was not being set as expected.

The 'output' variable is now set as local and the function is fixed to
set the CEPH_FAILURE variable accondingly.

Test-Plan:
  PASS: Create a script with 'output' variable set to random value and
        call the execute_ceph_cmd function with 'ceph health detail'
        command. Run the script with Ceph running. The CEPH_FAILURE
        must be different than 'true' and the 'output' variable must
        not change.

  PASS: Execute the same script as above, but with Ceph stopped. The
        CEPH_FAILURE variable must be equal 'true' and the 'output'
        variabel must not change.

Closes-bug: https://bugs.launchpad.net/starlingx/+bug/2083056

Signed-off-by: Felipe Sanches Zanoni <Felipe.SanchesZanoni@windriver.com>